+++ ui-lib/src/androidTest/java/co/electriccoin/zcash/ui/screen/scan/util/SettingsUtilTest.kt
@@ -1,10 +1,9 @@
 package co.electriccoin.zcash.ui.screen.scan.util
 
-import android.content.Context
 import android.content.Intent
 import android.provider.Settings
-import androidx.test.core.app.ApplicationProvider
 import androidx.test.filters.SmallTest
+import co.electriccoin.zcash.ui.test.getAppContext
 import org.junit.Test
 import kotlin.test.assertContains
 import kotlin.test.assertEquals
@@ -14,13 +13,13 @@ class SettingsUtilTest {
 
     companion object {
         val SETTINGS_URI = SettingsUtil.SETTINGS_URI_PREFIX +
-            ApplicationProvider.getApplicationContext<Context>().packageName
+            getAppContext().packageName
     }
 
     @Test
     @SmallTest
     fun check_intent_to_settings() {
-        val intent = SettingsUtil.newSettingsIntent(ApplicationProvider.getApplicationContext<Context>().packageName)
+        val intent = SettingsUtil.newSettingsIntent(getAppContext().packageName)
         assertNotNull(intent)
         assertEquals(Settings.ACTION_APPLICATION_DETAILS_SETTINGS, intent.action)
         assertContains(intent.categories, Intent.CATEGORY_DEFAULT)
