+++ ui-lib/src/androidTest/java/co/electriccoin/zcash/ui/screen/send/ext/WalletAddressExtTest.kt
@@ -1,11 +1,11 @@
 package co.electriccoin.zcash.ui.screen.send.ext
 
-import androidx.test.core.app.ApplicationProvider
 import androidx.test.filters.SmallTest
 import cash.z.ecc.sdk.fixture.WalletAddressFixture
+import co.electriccoin.zcash.ui.test.getAppContext
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
-import kotlin.test.Test
+import org.junit.Test
 import kotlin.test.assertEquals
 
 class WalletAddressExtTest {
@@ -14,7 +14,7 @@ class WalletAddressExtTest {
     @SmallTest
     @OptIn(ExperimentalCoroutinesApi::class)
     fun testAbbreviated() = runTest {
-        val actual = WalletAddressFixture.shieldedSapling().abbreviated(ApplicationProvider.getApplicationContext())
+        val actual = WalletAddressFixture.shieldedSapling().abbreviated(getAppContext())
 
         // TODO [#248]: The expected value should probably be reversed if the locale is RTL
         val expected = "ztest…rxnwg"
