+++ ui-lib/src/androidTest/java/co/electriccoin/zcash/ui/screen/update/util/PlayStoreUtilTest.kt
@@ -1,9 +1,8 @@
 package co.electriccoin.zcash.ui.screen.update.util
 
-import android.content.Context
 import android.content.Intent
-import androidx.test.core.app.ApplicationProvider
 import androidx.test.filters.SmallTest
+import co.electriccoin.zcash.ui.test.getAppContext
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNotNull
 import org.junit.Test
@@ -13,13 +12,13 @@ class PlayStoreUtilTest {
 
     companion object {
         val PLAY_STORE_URI = PlayStoreUtil.PLAY_STORE_APP_URI +
-            ApplicationProvider.getApplicationContext<Context>().packageName
+            getAppContext().packageName
     }
 
     @Test
     @SmallTest
     fun check_intent_for_store() {
-        val intent = PlayStoreUtil.newActivityIntent(ApplicationProvider.getApplicationContext())
+        val intent = PlayStoreUtil.newActivityIntent(getAppContext())
         assertNotNull(intent)
         assertEquals(intent.action, Intent.ACTION_VIEW)
         assertContains(PLAY_STORE_URI, intent.data.toString())
