+++ ui-lib/src/androidTest/java/co/electriccoin/zcash/ui/test/Global.kt
@@ -4,6 +4,8 @@ import android.content.Context
 import androidx.annotation.StringRes
 import androidx.test.core.app.ApplicationProvider
 
-fun getStringResource(@StringRes resId: Int) = ApplicationProvider.getApplicationContext<Context>().getString(resId)
+fun getAppContext(): Context = ApplicationProvider.getApplicationContext()
 
-fun getStringResourceWithArgs(@StringRes resId: Int, vararg formatArgs: String) = ApplicationProvider.getApplicationContext<Context>().getString(resId, *formatArgs)
+fun getStringResource(@StringRes resId: Int) = getAppContext().getString(resId)
+
+fun getStringResourceWithArgs(@StringRes resId: Int, vararg formatArgs: String) = getAppContext().getString(resId, *formatArgs)
