+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/fixture/WalletSnapshotFixture.kt
@@ -13,15 +13,18 @@ object WalletSnapshotFixture {
 
     val STATUS = Synchronizer.Status.SYNCED
     val PROGRESS = PercentDecimal.ZERO_PERCENT
+    val TRANSPARENT_BALANCE: WalletBalance = WalletBalance(Zatoshi(8), Zatoshi(1))
+    val ORCHARD_BALANCE: WalletBalance = WalletBalance(Zatoshi(5), Zatoshi(2))
+    val SAPLING_BALANCE: WalletBalance = WalletBalance(Zatoshi(4), Zatoshi(4))
 
     // Should fill in with non-empty values for better example values in tests and UI previews
     @Suppress("LongParameterList")
     fun new(
         status: Synchronizer.Status = STATUS,
         processorInfo: CompactBlockProcessor.ProcessorInfo = CompactBlockProcessor.ProcessorInfo(),
-        orchardBalance: WalletBalance = WalletBalance(Zatoshi(5), Zatoshi(2)),
-        saplingBalance: WalletBalance = WalletBalance(Zatoshi(4), Zatoshi(4)),
-        transparentBalance: WalletBalance = WalletBalance(Zatoshi(8), Zatoshi(1)),
+        orchardBalance: WalletBalance = ORCHARD_BALANCE,
+        saplingBalance: WalletBalance = SAPLING_BALANCE,
+        transparentBalance: WalletBalance = TRANSPARENT_BALANCE,
         pendingCount: Int = 0,
         progress: PercentDecimal = PROGRESS,
         synchronizerError: SynchronizerError? = null
